Update mapping method in the UuidMapFile class
The code that was used to migrate the id map files from `data-ids.csv`
to `idmap/data.csv` was throwing a rake error when adding a new
source and rebuilding the data with rake clean default, since there
was no @oldfile or @newfile. Moving the .read down to the
Rcsv.parse call seemed to fix the error.