fix (dashboard): ignore short lines
When reading the export file from Google Analytics, ignore any lines
that don't have enough fields.

Various parts of this file (particularly at the start and end) have an
entirely different format from the 'meat' that we care about, and
something that recently changed in the layout of this started breaking
the dashboard generation script. The easiest way to prevent this is to
simply ignore lines that don't look like they're in the correct format.