Mexico (Deputies) Revert problematic source official-LXIII.csv
There's a lot of other things going on in this PR, so this commit is
to revert only the source that's problematic, get a clean PR, and
merge it, then make another PR with the problematic source.

This commit recovers the official data for the LXIII term,
since it was identified as being the source of problems
with the import (names of some politicians ending in
LICENCIA).