EP_COUNTRY_REFRESH should work by slug
Previously we assumed we were working with a country in directory-name
format (e.g. South_Africa), but the Rebuilder is passing us slugs (i.e.
South-Africa). That's the more correct behaviour, so switch to assuming
that's what we're getting.

The implementation is a little bit clumsy at the moment — we can clean
things up substantially by making everything assume they'll be dealing
with EveryPolitician::Country or EveryPolitician::Legislature objects,
but that should happen as a separate change.