replace is_wikidata check
remove the `is_wikidata?` method, and replace it with a
`sources_of_type('wikidata')` check. This is bit more verbose, but now
that we _have_ the "of_type" check, it's better to use it than
cluttering up the public interface of every Source with a method that's
only true for one of them.