replace `is_memberships?` check
remove the `is_memberships?` method, and replace it with a
`sources_of_type('memberships')` check. This is bit more verbose, but
now
that we _have_ the "of_type" check, it's better to use it than
cluttering up the public interface of every Source with a method that's
only true for one of them.