Add a warnings system to `Source`s
Allow each source class to build up a list of unique warnings that can
be output in one shot later. This is a parallel system to the
'warn_once' mechanism in `combine_sources` so we can pull most of the
logic out of there over into the individual Source classes.