Switch to deploying everypolitician.org in after_success hook
It turns out that Travis' script based deployment support leaves a lot
to be desired. It appears to run the deployment script in a very
different environment to the rest of the build, which results in it
using a very outdated version of ruby and git that causes errors with
the deploy.

This change moves the logic for deploying everypolitician.org into the
`after_success` hook, which should be run in the same environment as the
rest of the build.