Unset BUNDLE_GEMFILE before trying to install viewer-sinatra
Travis sets a BUNDLE_GEMFILE environment variable pointing to the
`Gemfile` of the application that's being built, everypolitician-data in
this case. However we want to install viewer-sinatra so that we can
deploy everypolitician.org, so we need to unset BUNDLE_GEMFILE so that
`bundle install` picks up the correct `Gemfile` for viewer-sinatra.