Properly store newly generated UUIDs
id_map is a method here, not a hash, so setting to it silently fails.
Instead we need to inline it, and explicitly pass it when writing out.
(We could instead add proper setter methods on it, but this seems
simpler for now, as we still want to rewrite this whole area RSN)