MX: Remove (LICENCIA) from member names
Upstream (which has now also moved, so explicitly choose the correct
scraper rather than redirecting), some members are having a designation of
(LICENCIA) included as part of their names.

Blank that out as part of the SELECT, whilst we wait for it to be fixed in the
scraper:
  https://github.com/everypolitician-scrapers/mexico-diputados-2015/issues/3

NB: this is still being maintained in the `sort_name` column. I
considered removing it from that in the same manner, but actually I
suspect it's OK to leave that for now: it should have no ill effects in
that context, and lets us still keep a historical record of which people
have this set at which time.