TN: Archive vanishing members to manual file
Members were disappearing from upstream. Their IDs were being assigned to new members and
so it seemed that the original members were changing names as these IDs being paired to the
UUIDs of the original members.

This commit moves the original members and their ID/UUID pairings to a manual file and a separate
ID map file so that when the data is built with the new members, the new members will
receive their own UUIDs.