Add rake task to remap ids in idmap files from memberships
If the id of a politician changes, or a politician is given a UUID
because they didn't have one, we need to update our idmap files
to point the new ID at our UUID.

This adds a rake task to do that. You call it as:

```
bundle exec rake memberships:remap_ids[oldID,newID]
```

(some shells, e.g. zsh, may need to quote those square brackets)