Move elections logic from Rakefile to Source
This is a step towards making the build process have as little logic as
possible in the Rakefiles.

The general ideas is that each source will know how to return Popolo,
that will then all get stitched together. For now that stitching can
happen in sequence, as here, but the logic should later become generic
enough to happen in a loop over all the sources.