Remap a UUID to its new upstream id
If the id of a person changes upstream, we want to overwrite the old
id to point the UUID to the new id. This commit adds a method in the
id mapper class to do that. it checks both that the old id exists in the
file and that the new id doesn't already exist in said file.