JM: Only bring in images from jmelections source
The other information is useful for reconciliation, so let's leave that
behind as a 'reconciliation_query' entry (JSON doesn't allow comments…)
for later ease of use, but the only field we actually want to include is
the image.

I considered including the names as well, but most of them are either
the correct name with 'MP' on the end, or a mis-spelling that we
probably don't really want to add to the 'other names'. (Maybe if we had
a "known mis-spellings" section…)