Memoize the Position File JSON
When working with the Position file, don't keep reloading the JSON every
time we're asking for the Positions for a given Person. The build calls
this potentially thousands of time in sequence, for each person with
Wikidata.

Removing this speeds up this section of the build for Sweden (for
example) from over 15 seconds to 0.037 seconds.