Hong Kong: Archive the 5th Legislative Council
There has been a recent election in Hong Kong (on the 4th September
2016). The page that lists members for the 5th term (the one that's
ended) is still at the same URL, which includes the term dates, but
no longer links to the person pages, and consequently we can't find
IDs for them from the page any more.  So, it seems most sensible to
archive the data and idmap from the 5th session and then subsequently
update the scraper to point to the page for the 6th term.

This commit adds a manual source for the 5th Legislative Council, and
removes fetching data from the scraper. The scraper will need minor
changes, and then a similar source can be reinstated to fetch the
data for the 6th Legislative Council.