PH: Update instructions file to add scraper for term 17
A new block was added for the scraper for term 17 after the scraper
was moved to EP-scrapers in morph and its code changed to scrape
the new term. A new reconciliation/merge block was added inside it
since there are some members in the new term that are also in the
old term. The reconciliation is done by name because we can not
trust that the ids upstream are unique. Also, the source is not
exposing real ids, but building them for the urls.
(For example: ferrer-j is Ferrer, Jeffrey P. in term 16 but Ferrer, Juliet
Marie D. in term 17)