Allow Merging of Membership sources with same IDs
Previously we haven't needed to merge Membership sources with the same
ID, as those will almost always be coming from the same upstream source,
and will therefore be in the same Morph database, and can be fetched
together as a single 'source'.

Sometimes, however, we will have archived off a previous term to a
manual source file, and will want to merge this with a new source file
for the latest term from the same upstream source, with shared IDs.

In such a case — i.e. where we have an `instructions.json` stanza for a
second Membership source, and no `merge` details, we can simply re-use
any ID → UUID mapping we already know of for any of the people in this
source.