LA: Add separate import instruction for 2016 term
Both 2011 and 2016 terms are pulled in from the same source but as
we have no upstream IDs and members' names are spelled differently
across terms we need to reconcile the memberships manually.

Since we already have UUIDs for the 2011 members, we need to manually
reconcile any members that are also in 2016 with the existing
UUIDs to prevent duplicates.