LA: Import membership data by selecting 2011 term
Memberships do not have upstream IDs and matching by name
is precarious as spellings differ between terms.

The plan is to import terms to separate files and reconcile
new members with existing members using the reconciliation
interface.

The first step (this step) changes the existing import instruction
to specifically import the existing term.