Require csvlint
We want to be able to check that our CSV files are well formed. The
`csvlint` gem seems like our best option here. It's easy to use from the
command line or programmatically, and will let us evolve to schema
validation checks.

Unfortunately it requires us to also update our ruby version, due
to its dependence on a version of activesupport with a secondary
dependency on that.

We could potentially make this check only run if the ruby version is
high enough, but that seems quite difficult (particularly in terms of
generating a suitable Gemfile.lock)

However, as the code in this repo is mostly only run internally, or by
people who we suspect will have no difficuly using ruby 2.3, I think
it's OK to bump our minimum supported version.