Move csvlint from test to production
The rebuilder skips all gems that are in the `:test` group, so
attempting to run the csv linting task fails.

In theory we shouldn't _need_ to run that task at the point of the actual
build, but the extra check is useful, so move this to production setting
rather than test.