LA: Move 2011 data to own file
Changes the import instruction for term 2011 to store 2011 data
in file named: 'data-2011.csv'. Existing files (morph source and
id map) have been renamed.

This is a step towards importing 2016 to a seperate files.

The terms will need to be merged using the reconciler as we have no IDs
to use and name spellings are inconcistent.