Ensure "unknown" party has an ID
Often upstream sources don't provide us any party/group information for
Members. Our practice is to default such cases to Unknown unless we
explicitly know that they are Independent.

When doing that, we usually construct a pseudo-Party named "Unknown".
However, if the source explicitly sets the party field to "Unknown", we
use the Popolo Organization generated from that (in CSV-to-Popoolo)
instead.

When we generate such a Party ourselves, we ensure it has an ID
("party/_unknown"). However, when we re-use an existing "Unknown" party,
we need to also ensure that it has an ID too, otherwise we end up with
lots of broken references.