Don't fail on failure
The most common usage for this script is to run a build across every
country, for example to check if a change to the build process has
created any unexpected errors or diffs. If one legislature is failing,
it's better to continue to see where else has problems, than to fail a
relatively slow process halfway through.