Don't always assume same IDs
In https://github.com/everypolitician/everypolitician-data/pull/19223 we
introduced the ability to treat the incoming IDs from multiple sources
as being for the same person.

However, this isn't always going to be true, and we don't always want to
have to add a `merge` section when it's not (especially for a manual
file with a single member, where the reconciliation file would be empty)

So, allow a 'reuse_ids' instruction that if set to 'false', will fall
back on the old behaviour of assuming that the IDs are to map to
different UUIDs.