Switch new_tempfile to a `let` block
Occasionally the final test here would fail, saying there are zero
entries in the file rather than two.

This is very hard to replicate, so difficult to investigate, but we
suspect that it might be because nothing is holding a reference to the
Tempfile itself, and so sometimes the garbage collection kicks in and
deletes it.

Setting it up in a `let` block should keep it around long enough to not
disappear (if this _is_ the problem), and with no ill-effects even if
it's not.