MN: Amend instruction to import terms 2008 and 2012 only
We updated the scraper URL in the previous commit. Since the
new scraper pulls in term 2016 -- which is already set in terms.csv --
we want to prevent that term from being imported at the time being.
This is so that we can see how the new scraper affects the current
data.