Switch area-wikidata to Morph source
Rather than generating wikidata for Areas ourselves at build time,
delegate the Wikidata lookup to a Morph scraper, from which we can fetch
the data as normal, and then merge it at build time.

The old 'source' CSV, from which we decided which Wikidata items to look
up, now becomes a reconciliation file, to let us merge the morph data.