Fix Yajl double quote encoding error
This is a workaround for Yajl not correctly encoding objects that
contain strings that contain double quotes. I'm not entirely sure what
it is in Yajl that is causing this, I _think_ it's because they redefine
the `to_json` method in the `json_gem` file, but I haven't dug into it
any deeper than that.

This fixes the Ukraine error, which means that can now be built again.