Japan: handle changed format from scraper
The scraper now provides an ID, and some Japanese-language data.

We can re-reconcile later to use the given IDs, but for now we just skip
them. The language code for hiragana has changed upstream since this
data was produced, but rather than waiting for that to re-run, we rename
it now. We should be able to undo this part again in the next day or two.