Load local countries.json when rebuilding
A plain `EveryPolitician.countries` will fetch the remote version, which
is needless when we have one locally.

This is also important if we want to manually manipulate the local file
before rebuilding (e.g. here, where we're renaming a legislature, and
want to look in a different place for data than where the previous
countries.json pointed)