Australia (Senate): move to manual term file
This data was previously hardcoded into the scraper, which isn't really
the best place for it to live.

(NB, this is an identical file to the House of Representatives)

This data doesn't seem to be in Wikidata yet, so we can't yet generate
it from there. There are no invidivial pages for Australian Parliaments
there — only an overview page:
  https://en.wikipedia.org/wiki/Chronology_of_Australian_federal_parliaments