Extend the missing_data task to changing_sources
Rather than a single `missing_data:memberships` task, expand this to
report on rows either added or removed from a file. The interface is
now:

  * rake changing_sources:added_rows[source]
  * rake changing_sources:removed_rows[source]

There's a lot of duplication in the implementation of the commands, but
at this stage it's hard to tell how they might vary in the future, or
what other things might be added here, so I think it's best to avoid
premature abstraction.