Nigeria: update instructions.json to use the new morph scraper
The new version of the scraper also includes data on Senators, so this
commit also updates the SELECT statement in the instructions.json for the
'Representatives' legislature to only select Members of the House
of Representatives.

This looks for the value 'Hon' in the js_position column, which means
that the 'Position' box on that person's page says 'Member'.  There are
three representatives for whom this is true who don't also have 'House
of Representatives' in the 'chamber' column - they have 'NASS' in that
column. These are:

  http://www.nass.gov.ng/mp/profile/504
  http://www.nass.gov.ng/mp/profile/965
  http://www.nass.gov.ng/mp/profile/964

From Googling, I believe that these are Members of the House of
Representatives, but if that turns out to be correct, we can use the
'chamber' column instead.

The SELECT statement also excludes a person called 'Anonymous Anonymous'
found by the scraper, whom I don't think is real - that page is probably
present on the site for testing purposes or something similar.

In addition, it excludes people who have no area information - these are
typically duplicates or people whose election is being contested.