BR: Manually map member UUID to wikidata ID
The name of the member in Wikidata differs considerably from the
name in the incoming data. Both names do refer to the same person:

André Luiz de Carvalho Ribeiro is also known as André_Fufuca
https://en.wikipedia.org/wiki/André_Fufuca

I had to pair the IDs manually as the reconciler was not offering
the correct member as an option.