NG (Senate): use shorter-name field as name
The upstream scraper has changed 'name' to be the longer version, and
added a 'shorter_name' column. In a later change we'll switch to
importing both of these, but for now, simply alias the short_name as
name, so we can do that as an isolated change, once we've caught up with
other changes.