NG: Hard-code term ID in instructions
We want to switch the data coming from the upstream source, which is
being changed from year-based (2015) to using the actual term IDs (i.e.
8 for the current term).

Whilst that change is in process (https://github.com/everypolitician-scrapers/nigeria-national-assembly/pull/15),
hard-code the new value here, so we don't have to be completely in sync with that.