TC: Split 2012 archive source into elected and appointed members
The wikipedia source contains party and area data. As such, it
should be imported as a membership source.

A future PR will move the wikipedia source above the elected members
source. The wikipedia source type will be set to membership, whilst
the elected members source will be set to primary.

The wikipedia article lists elected members only. If this source
were to completely replace the original membership source, the four
appointed members would not be picked up. Splitting the archive
file into two means that the appointed member source can be imported
as `memberships` whilst the elected member source can be imported
as `person`.