Only deploy when countries.json changes
We don't want to deploy everypolitician.org unless there has been a
change to `countries.json`.

This uses the "TRAVIS_COMMIT_RANGE" environment variable to check if
there has been a change to `countries.json` in the commit being built.
If not then the deploy is skipped to avoid unnecessary spidering.