Always deploy when countries.json changes
Previously we were deploying when the only file that changed was
countries.json, but we also want to handle the case where countries.json
and a bunch of other files all change at once.

I've changed this check to use the `-z` operand [1] to test for zero
length output from the diff command, which indicates that
`countries.json` wasn't changed.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/test.html