Pakistan (Assembly): Add end date column to archive data
This is a step towards fixing: #24210

One member has two memberships in the current term. There is no end date for the first membership and so it overlaps the second.

The first membership is stored in the archive directory. As a step towards adding the end date for this membership, this commit adds an end_date column to the archived membership file.