Liechtenstein (Landtag): add new term information
Parline doesn't have any information other than the date of the
election.

http://www.landtag.li/agenda-en seems to imply that Feb 5th was the
start date. The last session of the previous term was in November, but
it's unclear if it dissolved then, or just didn't meet after that, so
I've set the date for now to be the day before the election.