TH: refresh data
We're getting a total of 250, which looks right according to the official
site... unless not everyone on there is current. It's very likely that
individuals have been changed in and out because of the nature of this
particular government right now.

So if we're losing any members (I think we are but I'll look at the PR
summary from the bot) we don't have end-dates for them.

Note: no reconciliation done here as I'm assuming that the photo IDs
are consistent and unique (which seems fairly likely); will check.