rake task to generate an area UUID mapping file
Largely a copy of the equivalent file for group UUIDs, except that
usually we have a reconciliation file here as area information comes
from a scraper rather than being generated directly. It may be possible
to abstract away the commonalities between these, but it's not clear yet
that that would be worthwhile, or if other differences will appear as we
start to roll this out everywhere.