Generate area + group IDs if missing
If a Membership source has an area or group column (after remapping the
names — e.g. this could also be called "constituency" or "party"), but
no equivalent area/group-id column, generate one based on slugifying the
name.

This is currently done much later in the process, via csv-to-popolo, but
moving it to this earlier stage gives us much control over remapping
these again (e.g. to UUIDs)

As a side-effect this also removes the needless "area/" or "group/"
prefix, which dates back to the very early days of OpenPoliticians, when
our process required IDs to be unique across the entire Popolo (i.e.
having a party and an area each with the same ID would be problematic)

As this hasn't been an issue for a long time, and these prefixes end up
being quite problematic (and need stripped off again in many places,
e.g. when comparing for equality), getting rid of them is very much a
bonus, though this will generate a lot of diffs initially.