Uganda (Parliament): remove OCD lookup
Part of moving away from OCD IDs everywhere, in favour of Wikidata IDs.

This introduces lots of new areas, where it's not totally clear, for
exmaple, whether two places with similar names in the 9th and 10th
Parliaments are actually the same (and thus we should update the name in
the spreadsheet), or there was a redistricting between the two terms (in
which case we shouldn't have been merging them previously).

I suspect a lot of them are the former, but the fuzzy lookup was always
problematic, so there are probably a mix of both. For now splitting
these up is simpler than trying to unravel them all, and hopefully we'll
end up with wikidata items for them all, and can resolve it properly
then.